requirements-local.txt
-r requirements.txt
django-debug-toolbar==4.0.0
factory-boy==3.3.0
Faker==25.3.0